[MODIFIED] Refactoring the repo
