+++ package.json
@@ -2,10 +2,11 @@
   "name": "persistent-typescript",
   "version": "1.0.0",
   "description": "An easy way to load & save variables",
-  "main": "index.js",
+  "main": "./build/Persistent.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "compile": "tsc --watch"
+    "compile": "tsc --watch",
+    "example": "tsc --watch -p ./example/"
   },
   "repository": {
     "type": "git",
